공격 이펙트 적용
및 플레이어 애니메이션 최적화

이펙트 오브젝트 풀링 적용

+++ Assets/0. Script/System/Soul/SoulData.cs
@@ -87,7 +87,7 @@ public class SoulData : ScriptableObject
         switch (effect.type)
         {
             case SoulEffectType.StatFlat:
-                return effect.flatValue;
+                return (int)effect.flatValue;
             case SoulEffectType.StatPercent:
                 return effect.percentValue;
             case SoulEffectType.IncreaseJumpNum:
